requirements.txt
requests
requests_ntlm
python-dotenv
PyYAML